Move irange test from test_valid_models to contrib
