+++ tests/contrib/autoguide/test_advi.py
@@ -49,6 +49,28 @@ def test_shapes(advi_class, trace_graph, enum_discrete):
     assert np.isfinite(loss), loss
 
 
+@pytest.mark.xfail(reason="irange is not yet supported")
+@pytest.mark.parametrize('advi_class', [ADVIDiagonalNormal, ADVIMultivariateNormal])
+def test_irange_smoke(advi_class):
+
+    def model():
+        x = pyro.sample("x", dist.Normal(0, 1))
+        assert x.shape == ()
+
+        for i in pyro.irange("irange", 3):
+            y = pyro.sample("y_{}".format(i), dist.Normal(0, 1).reshape([2, 1 + i, 2], extra_event_dims=3))
+            assert y.shape == (2, 1 + i, 2)
+
+        z = pyro.sample("z", dist.Normal(0, 1).reshape([2], extra_event_dims=1))
+        assert z.shape == (2,)
+
+        pyro.sample("obs", dist.Bernoulli(0.1), obs=torch.tensor(0))
+
+    advi = advi_class(model)
+    infer = SVI(advi.model, advi.guide, Adam({"lr": 1e-6}), "ELBO")
+    infer.step()
+
+
 @pytest.mark.parametrize("advi_class", [ADVIMultivariateNormal, ADVIDiagonalNormal])
 def test_median(advi_class):
